feat: basic interfaces with itable working

+++ runtime/abi.zig
@@ -14,7 +14,9 @@ pub const ZeusType = enum(u8) { _i8 = 0, _i16 = 1, _i32 = 2, _i64 = 3, _f32 = 4,
 pub const ZeusObjectTypeInfo = extern struct {
     // unique id given to every class in the program
     // primordial classes with have a known fixed id
-    object_type_id: u8,
+    // i32-wide to match codegen's ZeusObjectTypeInfo (interface dispatch keys on this id, so
+    // the class-id space must not be capped at 255).
+    object_type_id: u32,
     object_type: ZeusObjectType,
     // if the object is an array, this is the type of the elements in the array
     // if it is not an array then this will have _null type
